+++ 30_sources/SRC_Zupfnoter/src/controller_command_definitions.rb
@@ -59,6 +59,26 @@ class Controller
       end
     end
 
+
+    @commands.add_command(:autofold) do |c|
+      c.undoable = false
+      c.set_help { "set automatic folding of config in editor #{c.parameter_help(0)}" }
+      c.add_parameter(:value, :boolean) do |parameter|
+        parameter.set_default { 'true' }
+        parameter.set_help { "true | false" }
+      end
+      c.as_action do |args|
+        value = {'true' => true, 'false' => false}[args[:value]]
+        if value
+          @editor.autofold = true
+          @editor.fold_all
+        else
+          @editor.autofold = false
+          @editor.unfold_all
+        end
+      end
+    end
+
     @commands.add_command(:autorefresh) do |c|
       c.undoable = false
       c.set_help { "turnon autorefresh" }